Make -arXiv-id work for more URLs
Now, all URLs are supported, not only maths one—there is more on the
arXiv, after all.  Additionally, -pdf-link is a little bit more
forgiving in what it accepts as an input.

As an added benefit, this means that we can now call -get-details with a
PDF as well, which transcends all the way up to arXiv-citation and -gui.